// types.d.ts
export interface Genre {
    id: number;
    name: string;
}

export interface Movie {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date?: string;
    title: string;
    video: boolean;
    media_type?: string;
    first_air_date: string;
    name: string;
    original_country: string[];
    vote_average: number;
    vote_count: number;
}

export interface Element {
    type: 'Bloopers' | 'Featurette' | 'Behind The Scenes' | 'Clip' | 'Trailer' | 'Teaser';
}